+++ public/rss-all.xml
@@ -4,7 +4,7 @@
         <description><![CDATA[Amir R Muntasser's Firehose]]></description>
         <link>https://arkm.xyz</link>
         <generator>RSS for Node</generator>
-        <lastBuildDate>Sat, 17 Apr 2021 07:05:10 GMT</lastBuildDate>
+        <lastBuildDate>Sat, 17 Apr 2021 07:50:21 GMT</lastBuildDate>
         <item>
             <title><![CDATA[Creating an RSS feed with Next.js and Next-MDX]]></title>
             <link>https://arkm.xyz/posts/create-an-rss-feed-nextjs-mdx-blog</link>
