+++ public/rss-weaknotes.xml
@@ -4,7 +4,7 @@
         <description><![CDATA[Amir R Muntasser's Weaknotes]]></description>
         <link>https://arkm.xyz/weaknotes</link>
         <generator>RSS for Node</generator>
-        <lastBuildDate>Sat, 17 Apr 2021 07:05:10 GMT</lastBuildDate>
+        <lastBuildDate>Sat, 17 Apr 2021 07:50:21 GMT</lastBuildDate>
         <item>
             <title><![CDATA[Week 1]]></title>
             <link>https://arkm.xyz/posts/week-1</link>
