+++ src/styles/Panels.module.css
